ckan-backend-dev/src/ckanext-wri/dev-requirements.txt
pytest-ckan
ckanapi==4.7
